added docstrings + sphinx gallery
