--- docs/source/deepinv.examples.pnp.rst
@@ -1,13 +1,0 @@
-Plug-and-Play
--------------------------------------
-These examples illustrate how to load pretrained denoisers and use the plug-and-play framework for solving inverse problems.
-
-Deblurring
-^^^^^^^^^^^^^^^^^^^^^^^^^^
-.. literalinclude:: ../../examples/demo_pnp_deblur.py
-   :language: python
-
-Compressed Sensing
-^^^^^^^^^^^^^^^^^^^^^^^^^^
-.. literalinclude:: ../../examples/demo_pnp_cs.py
-   :language: python
