--- docs/source/deepinv.examples.rst
@@ -1,13 +1,0 @@
-Examples
-=================================
-
-Add Google Colab link if possible.
-
-.. toctree::
-   :maxdepth: 2
-   :caption: Contents:
-
-   deepinv.examples.basics
-   deepinv.examples.pnp
-   deepinv.examples.unrolled
-   deepinv.examples.sampling
