--- docs/source/deepinv.examples.sampling.rst
@@ -1,8 +1,0 @@
-Sampling
--------------------------------------
-These code snippets show you how to use sampling algorithms to quantify uncertainty of a reconstruction.
-
-MCMC sampling
-^^^^^^^^^^^^^^^^^^^^^^^^^^
-.. literalinclude:: ../../examples/demo_sampling.py
-   :language: python
