--- docs/source/deepinv.examples.unrolled.rst
@@ -1,3 +1,0 @@
-Unrolling
--------------------------------------
-Work in progress...
