+++ examples/README.rst
@@ -1,0 +1,3 @@
+Examples
+==================
+
