+++ examples/basics/demo_train_inpainting.py
@@ -1,3 +1,9 @@
+r"""
+Training a reconstruction network.
+====================================================================================================
+
+"""
+
 import deepinv as dinv
 from torch.utils.data import DataLoader
 import torch
