+++ examples/plug-and-play/demo_PnP_DPIR_deblur.py
@@ -1,26 +1,27 @@
 """
-Implementation of the DPIR method for Plug-and-Play image deblurring. 
+Image deblurring with DPIR PnP.
+=================================================================================
 
-Zhang, K., Zuo, W., Gu, S., & Zhang, L. (2017). 
-Learning deep CNN denoiser prior for image restoration. 
+Zhang, K., Zuo, W., Gu, S., & Zhang, L. (2017).
+Learning deep CNN denoiser prior for image restoration.
 In Proceedings of the IEEE conference on computer vision and pattern recognition (pp. 3929-3938).
 """
+
 import numpy as np
 import deepinv as dinv
-import hdf5storage
 from pathlib import Path
 import torch
 from torch.utils.data import DataLoader
 from deepinv.models.denoiser import Denoiser
 from deepinv.optim.data_fidelity import L2
-from deepinv.optim.optimizers import optimbuilder
+from deepinv.optim import optim_builder
 from deepinv.training_utils import test
 from torchvision import datasets, transforms
 from deepinv.utils.parameters import get_DPIR_params
 from deepinv.utils.demo import get_git_root, download_dataset, download_degradation
 
 # Setup paths for data loading and results.
-BASE_DIR = Path(get_git_root())
+BASE_DIR = Path(".")
 ORIGINAL_DATA_DIR = BASE_DIR / "datasets"
 DATA_DIR = BASE_DIR / "measurements"
 RESULTS_DIR = BASE_DIR / "results"
@@ -74,9 +75,12 @@ kernel_index = 1  # which kernel to chose among the 8 motion kernels from 'Levin
 kernel_path = DEG_DIR / "kernels" / "Levin09.mat"
 if not kernel_path.exists():
     download_degradation("Levin09.mat", DEG_DIR / "kernels")
-kernels = hdf5storage.loadmat(str(kernel_path))["kernels"]
-filter_np = kernels[0, kernel_index].astype(np.float64)
+
+f = h5py.File(str(kernel_path), 'r')  # hdf5storage.loadmat(str(kernel_path))["kernels"]
+filter_np = f.get('kernels')[kernel_index]
+filter_np = np.array(filter_np).astype(np.float64)
 filter_torch = torch.from_numpy(filter_np).unsqueeze(0).unsqueeze(0)
+
 # The BlurFFT instance from physics enables to compute efficently backward operators with Fourier transform.
 p = dinv.physics.BlurFFT(
     img_size=(n_channels, img_size, img_size),
