+++ examples/plug-and-play/demo_custom_prior.py
@@ -1,9 +1,7 @@
 """
-Implementation of the GSPnP method for RED image super-resolution. 
+Image deblurring with explicit deep prior
+==============================================
 
-Hurault, S., Leclaire, A., & Papadakis, N. 
-Gradient Step Denoiser for convergent Plug-and-Play. 
-In International Conference on Learning Representations.
 """
 
 import numpy as np
@@ -12,13 +10,21 @@ import hdf5storage
 from pathlib import Path
 import torch
 from torch.utils.data import DataLoader
-from deepinv.models.denoiser import ScoreDenoiser
+import torch.nn as nn
 from deepinv.optim.data_fidelity import L2
-from deepinv.optim.optimizers import optimbuilder
+from deepinv.optim.optimizers import Optim
 from deepinv.training_utils import test
 from torchvision import datasets, transforms
-from deepinv.utils.parameters import get_GSPnP_params
 from deepinv.utils.demo import get_git_root, download_dataset, download_degradation
+
+
+# Create a nn.Module class to parametrize the custom prior
+class L2Prior(nn.Module):
+    def __init__(self, prior_params=None):
+        super(L2Prior, self).__init__()
+
+    def forward(self, x, g_param):
+        return torch.norm(x.view(x.shape[0], -1), p=2, dim=-1)
 
 
 # Setup paths for data loading, results and checkpoints.
@@ -27,25 +33,22 @@ ORIGINAL_DATA_DIR = BASE_DIR / "datasets"
 DATA_DIR = BASE_DIR / "measurements"
 RESULTS_DIR = BASE_DIR / "results"
 DEG_DIR = BASE_DIR / "degradations"
-CKPT_DIR = BASE_DIR / "ckpts"
 
-# Set the global random seed from pytorch to ensure reprod
-# ucibility of the example.
+# Set the global random seed from pytorch to ensure reproducibility of the example.
 torch.manual_seed(0)
 
-
-# Setup the variable to fetch dataset and operators.
-method = "GSPnP"
-denoiser_name = "gsdrunet"
-dataset_name = "CBSD68"
-operation = "super-resolution"
+# Set up the variable to fetch dataset and operators.
+method = "L2_prior"
+dataset_name = "set3c"
+operation = "deblur"
 dataset_path = ORIGINAL_DATA_DIR / dataset_name
 if not dataset_path.exists():
     download_dataset(dataset_name, ORIGINAL_DATA_DIR)
 measurement_dir = DATA_DIR / dataset_name / operation
 
 
-# Use parallel dataloader if using a GPU to fasten training, otherwise, as all computes are on CPU, use synchronous dataloading.
+# Use parallel dataloader if using a GPU to fasten training, otherwise, as all computes are on CPU, use synchronous
+# data loading.
 num_workers = 4 if torch.cuda.is_available() else 0
 
 
@@ -56,12 +59,17 @@ noise_level_img = 0.03  # Gaussian Noise standart deviation for the degradation
 img_size = 256
 n_channels = 3  # 3 for color images, 1 for gray-scale images
 early_stop = True  # Stop algorithm when convergence criteria is reached
-crit_conv = "cost"  # Convergence is reached when the difference of cost function between consecutive iterates is smaller than thres_conv
+crit_conv = "cost"  # Convergence is reached when the difference of cost function between consecutive iterates is
+# smaller than thres_conv
 thres_conv = 1e-5
 backtracking = True  # use backtraking to automatically adjust the stepsize
 factor = 2  # down-sampling factor
 use_bicubic_init = False  # Use bicobic interpolation to initialize the algorithm
-
+max_iter = 500  # Maximum number of iterations
+stepsize_prox_inter = 1.0  # Stepsize used for gradient descent calculation of the prox of the custom prior.
+max_iter_prox_inter = 50  # Maximum number of iterations for gradient descent calculation of the prox of the custom
+# prior.
+tol_prox_inter = 1e-3  # Convergence criteria for gradient descent calculation of the prox of the custom prior.
 
 # Logging parameters
 verbose = True
@@ -71,50 +79,40 @@ plot_images = True  # plot results
 save_images = True  # save images in RESULTS_DIR
 
 
-# Generate the degradation operator.
-kernel_index = 2  # which kernel to chose
-kernel_path = DEG_DIR / "kernels" / "kernels_12.mat"
+# Generate a motion blur operator.
+kernel_index = 1  # which kernel to chose among the 8 motion kernels from 'Levin09.mat'
+kernel_path = DEG_DIR / "kernels" / "Levin09.mat"
 if not kernel_path.exists():
-    download_degradation("kernels_12.mat", DEG_DIR / "kernels")
+    download_degradation("Levin09.mat", DEG_DIR / "kernels")
 kernels = hdf5storage.loadmat(str(kernel_path))["kernels"]
 filter_np = kernels[0, kernel_index].astype(np.float64)
 filter_torch = torch.from_numpy(filter_np).unsqueeze(0).unsqueeze(0)
-p = dinv.physics.Downsampling(
+
+# The BlurFFT instance from physics enables to compute efficiently backward operators with Fourier transform.
+p = dinv.physics.BlurFFT(
     img_size=(n_channels, img_size, img_size),
-    factor=factor,
     filter=filter_torch,
     device=dinv.device,
     noise_model=dinv.physics.GaussianNoise(sigma=noise_level_img),
 )
 
-# load specific parameters for GSPnP
-lamb, sigma_denoiser, stepsize, max_iter = get_GSPnP_params(
-    operation, noise_level_img, kernel_index
-)
-
-params_algo = {"stepsize": stepsize, "g_param": sigma_denoiser, "lambda": lamb}
 
 # Select the data fidelity term
 data_fidelity = L2()
 
-# Specify the Denoising prior
-ckpt_path = CKPT_DIR / "gsdrunet.ckpt"
-model_spec = {
-    "name": denoiser_name,
-    "args": {
-        "in_channels": n_channels + 1,
-        "out_channels": n_channels,
-        "pretrained": str(ckpt_path) if ckpt_path.exists() else "download",
-        "train": False,
-        "device": dinv.device,
-    },
-}
-# The prior g needs to be a dictionary with specified "g" and/or proximal operator "prox_g" and/or gradient "grad_g".
-# For RED image restoration, the denoiser replaces "grad_g".
 
-denoiser = ScoreDenoiser(model_spec, sigma_normalize=False)
-prior = {"grad_g": denoiser, "g": denoiser.denoiser.potential}
+# Specify the custom prior
+prior = {"g": L2Prior()}
 
+
+# Specific parameters for restoration with the given prior (Note that these parameters have not been optimized here)
+params_algo = {"stepsize": 1, "g_param": 1.0, "lamb": 1}
+
+
+# Design the cost function that is minimized by the algorithm.
+F_fn = lambda x, cur_params, y, physics: params_algo["lamb"][0] * data_fidelity.f(
+    physics.A(x), y
+) + prior["g"][0](x, cur_params["g_param"])
 
 # Generate a dataset in a HDF5 folder in "{dir}/dinv_dataset0.h5'" and load it.
 val_transform = transforms.Compose(
@@ -135,17 +133,8 @@ dataloader = DataLoader(
     dataset, batch_size=batch_size, num_workers=num_workers, shuffle=False
 )
 
-# By default the algorithm is initialized with the adjoint of the degradation matrix applied to the degraded image.
-# For custom initialization, we need to write a a function of the degraded image.
-if use_bicubic_init:
-    custom_init = lambda y: torch.nn.functional.interpolate(
-        y, scale_factor=factor, mode="bicubic"
-    )
-else:
-    custom_init = None
-
-# instanciate the algorithm class to solve the IP problem.
-model = optimbuilder(
+# Instantiate the algorithm class to solve the IP problem.
+model = Optim(
     algo_name="PGD",
     prior=prior,
     g_first=True,
@@ -156,10 +145,9 @@ model = optimbuilder(
     crit_conv=crit_conv,
     thres_conv=thres_conv,
     backtracking=backtracking,
-    return_dual=True,
+    F_fn=F_fn,
     verbose=verbose,
     return_metrics=plot_metrics,
-    custom_init=custom_init,
 )
